# doc/swagger.yaml
openapi: 3.0.3
info:
  title: T.W.I.N API - Ticket Workflow Integration Network
  description: |
    T.W.I.N (Ticket Workflow Integration Network) is a service designed to handle Jira webhooks and synchronize ticket workflows.

    This API provides endpoints for:
    - Receiving and processing Jira webhooks
    - Health monitoring
    - Testing connectivity

    **Author**: Pandor
    **Organization**: Aincrad-Flux
  version: 1.0.0
  contact:
    name: Aincrad-Flux
    url: https://github.com/Aincrad-Flux/TWIN
  license:
    name: ISC
    url: https://opensource.org/licenses/ISC
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://devgoblin.me
    description: Production server
paths:
...
  /api/config/config:
    get:
      summary: Get application configuration
      description: Retrieves the current application configuration including environment settings, version, and database configuration.
      tags:
        - Configuration
      security:
        - AdminApiKey: []
      responses:
        '200':
          description: Configuration retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigResponse'
              example:
                status: "success"
                data:
                  env: "development"
                  port: "3000"
                  version: "1.0.0"
                  logLevel: "info"
                  dbHost: "localhost"
                  dbPort: "5432"
                  dbName: "twin_db"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/logs/logs:
    get:
      summary: Get list of all log files
      description: Retrieves a list of all available log files with their metadata including size, modification date, and type.
      tags:
        - Logs
      security:
        - AdminApiKey: []
      responses:
        '200':
          description: Log files list retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogFilesResponse'
              example:
                status: "success"
                data:
                  files:
                    - name: "combined.log"
                      size: 15420
                      sizeFormatted: "15.06 KB"
                      modified: "2025-08-06T10:30:00.000Z"
                      type: "log"
                    - name: "2025-08-06/webhook-test-2025-08-06T12-30-40-375Z.json"
                      size: 2048
                      sizeFormatted: "2.00 KB"
                      modified: "2025-08-06T12:30:40.000Z"
                      type: "webhook"
                  total: 15
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/logs/logs/stats:
    get:
      summary: Get log statistics
      description: Retrieves statistics about log files including total count, size, and breakdown by type.
      tags:
        - Logs
      security:
        - AdminApiKey: []
      responses:
        '200':
          description: Log statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogStatsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/logs/logs/errors:
    get:
      summary: Get recent error logs
      description: Retrieves recent error logs from the error.log file.
      tags:
        - Logs
      security:
        - AdminApiKey: []
      parameters:
        - name: lines
          in: query
          description: Number of recent lines to retrieve
          schema:
            type: integer
            default: 50
          example: 100
      responses:
        '200':
          description: Error logs retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogContentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/logs/logs/combined:
    get:
      summary: Get recent combined logs
      description: Retrieves recent combined logs from the combined.log file.
      tags:
        - Logs
      security:
        - AdminApiKey: []
      parameters:
        - name: lines
          in: query
          description: Number of recent lines to retrieve
          schema:
            type: integer
            default: 100
          example: 200
      responses:
        '200':
          description: Combined logs retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogContentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/logs/logs/search:
    get:
      summary: Search across log files
      description: Searches for a specific term across all log files with optional filtering.
      tags:
        - Logs
      security:
        - AdminApiKey: []
      parameters:
        - name: q
          in: query
          required: true
          description: Search query
          schema:
            type: string
          example: "webhook"
        - name: file
          in: query
          description: File pattern to filter search
          schema:
            type: string
          example: "combined"
        - name: limit
          in: query
          description: Maximum number of results
          schema:
            type: integer
            default: 100
          example: 50
        - name: case_sensitive
          in: query
          description: Case sensitive search
          schema:
            type: boolean
            default: false
          example: false
      responses:
        '200':
          description: Search completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogSearchResponse'
        '400':
          description: Missing required query parameter
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/logs/logs/webhooks/{date}:
    get:
      summary: Get webhook logs for specific date
      description: Retrieves webhook logs for a specific date in YYYY-MM-DD format.
      tags:
        - Logs
      security:
        - AdminApiKey: []
      parameters:
        - name: date
          in: path
          required: true
          description: Date in YYYY-MM-DD format
          schema:
            type: string
            pattern: '^\d{4}-\d{2}-\d{2}$'
          example: "2025-08-06"
      responses:
        '200':
          description: Webhook logs for date retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookLogsResponse'
        '400':
          description: Invalid date format
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/logs/logs/webhooks/recent:
    get:
      summary: Get recent webhook logs
      description: Retrieves webhook logs from the last 24 hours.
      tags:
        - Logs
      security:
        - AdminApiKey: []
      parameters:
        - name: limit
          in: query
          description: Maximum number of files to return
          schema:
            type: integer
            default: 10
          example: 20
      responses:
        '200':
          description: Recent webhook logs retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookLogsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/logs/logs/cleanup:
    delete:
      summary: Clean up old log files
      description: Deletes log files older than the specified number of days.
      tags:
        - Logs
      security:
        - AdminApiKey: []
      parameters:
        - name: days
          in: query
          description: Number of days to keep (delete older files)
          schema:
            type: integer
            default: 30
          example: 7
      responses:
        '200':
          description: Log cleanup completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CleanupResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
components:
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - timestamp
      properties:
        error:
          type: string
          description: Error message
          example: "Processing error"
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the error occurred
          example: "2025-08-06T10:30:00.000Z"
...
    ConfigResponse:
      type: object
      required:
        - status
        - data
      properties:
        status:
          type: string
          example: "success"
        data:
          type: object
          properties:
            env:
              type: string
              example: "development"
            port:
              type: string
              example: "3000"
            version:
              type: string
              example: "1.0.0"
            logLevel:
              type: string
              example: "info"
            dbHost:
              type: string
              example: "localhost"
            dbPort:
              type: string
              example: "5432"
            dbName:
              type: string
              example: "twin_db"
...
    LogFile:
      type: object
      properties:
        name:
          type: string
          example: "combined.log"
        size:
          type: integer
          example: 15420
        sizeFormatted:
          type: string
          example: "15.06 KB"
        modified:
          type: string
          format: date-time
          example: "2025-08-06T10:30:00.000Z"
        type:
          type: string
          enum: ["log", "webhook", "json"]
          example: "log"
    LogFilesResponse:
      type: object
      required:
        - status
        - data
      properties:
        status:
          type: string
          example: "success"
        data:
          type: object
          properties:
            files:
              type: array
              items:
                $ref: '#/components/schemas/LogFile'
            total:
              type: integer
              example: 15
    LogStatsResponse:
      type: object
      required:
        - status
        - data
      properties:
        status:
          type: string
          example: "success"
        data:
          type: object
          properties:
            totalFiles:
              type: integer
              example: 15
            totalSize:
              type: integer
              example: 245760
            totalSizeFormatted:
              type: string
              example: "240.00 KB"
            byType:
              type: object
              additionalProperties:
                type: object
                properties:
                  count:
                    type: integer
                  size:
                    type: integer
                  sizeFormatted:
                    type: string
            recentActivity:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                  modified:
                    type: string
                    format: date-time
                  size:
                    type: integer
    LogContentResponse:
      type: object
      required:
        - status
        - data
      properties:
        status:
          type: string
          example: "success"
        data:
          oneOf:
            - type: object
              properties:
                filename:
                  type: string
                totalLines:
                  type: integer
                returnedLines:
                  type: integer
                content:
                  type: array
                  items:
                    type: string
            - type: object
              description: JSON log file content
              additionalProperties: true
    LogSearchResponse:
      type: object
      required:
        - status
        - data
      properties:
        status:
          type: string
          example: "success"
        data:
          type: object
          properties:
            query:
              type: string
              example: "webhook"
            totalResults:
              type: integer
              example: 25
            results:
              type: array
              items:
                type: object
                properties:
                  file:
                    type: string
                    example: "combined.log"
                  line:
                    type: integer
                    example: 42
                  content:
                    type: string
                    example: "Webhook request data saved"
                  timestamp:
                    type: string
                    format: date-time
    WebhookLogsResponse:
      type: object
      required:
        - status
        - data
      properties:
        status:
          type: string
          example: "success"
        data:
          type: object
          properties:
            date:
              type: string
              example: "2025-08-06"
            files:
              type: array
              items:
                $ref: '#/components/schemas/LogFile'
            total:
              type: integer
              example: 5
            period:
              type: string
              example: "last 24 hours"
    CleanupResponse:
      type: object
      required:
        - status
        - message
        - data
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "Log cleanup completed successfully"
        data:
          type: object
          properties:
            deletedCount:
              type: integer
              example: 5
            deletedSize:
              type: integer
              example: 102400
            deletedSizeFormatted:
              type: string
              example: "100.00 KB"
            remainingFiles:
              type: integer
              example: 10
...
  responses:
    Unauthorized:
      description: Unauthorized - API key missing
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "API key required"
            timestamp: "2025-08-06T10:30:00.000Z"
    Forbidden:
      description: Forbidden - Invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Invalid API key"
            timestamp: "2025-08-06T10:30:00.000Z"
...
    ServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Internal server error"
            timestamp: "2025-08-06T10:30:00.000Z"
  securitySchemes:
    JiraWebhook:
      type: apiKey
      in: header
      name: User-Agent
      description: |
        Jira webhook authentication is validated by checking the User-Agent header.
        The service expects requests from Atlassian services (User-Agent containing "Atlassian").

        Additional security validations may include:
        - IP address whitelist validation
        - HMAC signature verification (when configured)
        - Rate limiting

    AdminApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Admin API key for accessing configuration and logs management endpoints.

        **Header format**: `X-API-Key: your-admin-api-key`

        **Alternative format**: `Authorization: Bearer your-admin-api-key`

        This key is required for all administrative operations including:
        - Configuration management
        - Environment variables management
        - Logs access and management

        The admin API key is protected and cannot be retrieved or modified via the API.
...